added usage of card functions to main
